+++ lib/common/Memory/VirtualAllocWrapper.h
@@ -35,16 +35,16 @@ class PreReservedVirtualAllocWrapper
 {
 public:
 #if _M_IX86_OR_ARM32
-    static const size_t PreReservedAllocationSegmentCount = 256; // (256 * 64K) == 16 MB, if 64k is the AllocationGranularity
+    static const uint PreReservedAllocationSegmentCount = 256; // (256 * 64K) == 16 MB, if 64k is the AllocationGranularity
 #else // _M_X64_OR_ARM64
-    static const size_t PreReservedAllocationSegmentCount = 4096; //(4096 * 64K) == 256MB, if 64k is the AllocationGranularity
+    static const uint PreReservedAllocationSegmentCount = 4096; //(4096 * 64K) == 256MB, if 64k is the AllocationGranularity
 #endif
     
 public:
     PreReservedVirtualAllocWrapper();
     BOOL Shutdown();
     LPVOID      Alloc(LPVOID lpAddress, size_t dwSize, DWORD allocationType, DWORD protectFlags, bool isCustomHeapAllocation = false);
-    BOOL        Free(LPVOID lpAddress, size_t dwSize, DWORD dwFreeType);
+    BOOL        Free(LPVOID lpAddress,  size_t dwSize, DWORD dwFreeType);
     bool        IsPreReservedRegionPresent();
     bool        IsInRange(void * address);
     LPVOID      GetPreReservedStartAddress();
